chore: send blacklist notification to collab (#1430)
## Problem
Currently, collaborators only receive notifications on Pipe errors, but do not receive notifications for blacklisted emails.

## Solution
Also send notifications for blacklisted emails to collaborators.

## What changed?
* Refactored the logic for retrieving collaborators into a shared helper function
* Reuse logic for notifying collaborators for blacklisted emails

## Tests
- [ ] Error emails should still be sent correctly based on the error configuration
  - [ ] First error per day
    - [ ] Owner only
    - [ ] Owner + Editors
    - [ ] Owner + Editors + Viewers
  - [ ] All errors
    - [ ] Owner only
    - [ ] Owner + Editors
    - [ ] Owner + Editors + Viewers
- [ ] Blacklist notification emails should be sent correctly based on the error configuration
  - [ ] Owner only
  - [ ] Owner + Editors
  - [ ] Owners + Editors + Viewers
